requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.6
pywebview
Markdown==3.6
packaging==24.0
pillow==10.2.0
requests==2.31.0
soupsieve==2.5
tkinterhtml
urllib3==2.2.1
cefpython3
pygtk 
webkit
